Refactoring of mark buttons code and support for button "activation"
Now the buttons are connected to an IBOutlet collection in a custom UIView subclass: MarkButtonsArrayView.
The DayInfoView now has methods for activating the buttons and for assigning them to templates.

ShiftTypeMarkView has been removed since no more used.
DayInfoView has been moved to the source code directory -- previously it was misplaced in the root directory.