Fixed initialization bug and improved dayInfo animation
Now the animation occurs only when the shifts are changed.
The bug fix concerns the problem of how to handle the first launch of
the application or the case where the user does not grant the application
rights to access the calendars. The fix currently make the application
work, but it requires the user to relaunch the application.