Improved visualisation of cells
In this new version the dates before today are dimmed and the date outside the current month are even more dimmed.
This allows for a better visualisation of the current position inside the month in case the app is used to look at what happens
in the following few days.