Improved color scheme for inactive shifts, removed non-month dates
The shift buttons have now black labels when the shift is not set and white labels when they are set -- this improve the contrast with the buttons background and looks better.
Date cells  for dates not in the current month are now invisible - they were untappable anyway. This removes visual clutter and does not induce the user to think that they behave like the other cells.