Cleaner handling of bad UserActivity descriptions
Sometimes, due for instance in a change in how we index things, objects returned to the app when the user tap on a search result
can be unreadable using the current conventions. Instead of making the app crash the current implementation reset the index and
proceeds without selecting the proper date.