Created Workspace and TShiftsFramework
Big refactoring. Now the app is divided into three instead of two pieces (it was main app and extension). The main app has been split into a separate framework and the UI code (not sure the split is so clean right now).

The split allows for sharing code between the app and the extension and will help big ways into improving the extension.